requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
alembic==1.13.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
langchain==0.3.9
langchain-openai==0.2.10
langchain-community==0.3.9
chromadb==0.5.23
langfuse>=2.55.0
openai>=1.58.0
google-generativeai>=0.3.0
websockets>=12.0
